% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantport-package.R
\docType{package}
\name{quantport-package}
\alias{quantport}
\alias{quantport-package}
\title{quantport: Portfolio Optimization Routines}
\description{
Portfolio Optimization using Quadratic Programming and Sequential Programming
}
\author{
\strong{Maintainer}: Thanh Nguyen Minh \email{thanhuwe8@example.com}

}
\keyword{internal}
