% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showEfficient.R
\name{showEfficient}
\alias{showEfficient}
\title{PlotEfficient Frontier}
\usage{
showEfficient(ret, covmat, short, rf = 0, freq, simpoints, assetpoints)
}
\arguments{
\item{assetpoints}{}
}
\description{
PlotEfficient Frontier
}
