% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantport-package.R
\docType{package}
\name{quantport-package}
\alias{quantport}
\alias{quantport-package}
\title{quantport: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thanhuwe8/quantport}
  \item Report bugs at \url{https://github.com/thanhuwe8/quantport/issues}
}

}
\keyword{internal}
