% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showEfficient.R
\name{ShowEfficient}
\alias{ShowEfficient}
\title{PlotEfficient Frontier}
\usage{
ShowEfficient(ret, covmat, short, rf = 0, freq, simpoints, assetpoints)
}
\arguments{
\item{assetpoints}{}
}
\description{
PlotEfficient Frontier
}
